requirements.txt
fastapi
uvicorn
llama-index-core
llama-index-readers-file
llama-index-embeddings-ollama
llama-index-vector-stores-lancedb
lancedb
python-multipart
pydantic-settings
structlog
pyyaml
sentence-transformers
pytest
ruff
mypy
